Relocate Drv.
